# documentation/swagger.yaml
swagger: "2.0"
info:
  title: "Schpoop"
  description: "Platform for memes."
  version: "1.0.0"
host: "TBA"
...
paths:
...
  /users:
    post:
      tags:
        - "users"
      summary: "Register a new user"
      consumes:
        - "application/json"
      parameters:
        - in: "body"
          name: "body"
          description: "User to be registered"
          required: true
          schema:
            $ref: "#/definitions/User"
      responses:
        201:
          description: "User successfully registered"
        400:
          description: "Bad Request, see error message"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
    get:
      tags:
        - "users"
      summary: "Get a list of usernames"
      produces:
        - "application/json"
      parameters:
        - name: "username"
          in: "query"
          description: "Search query for usernames"
          required: false
          type: "string"
        - name: "pageSize"
          in: "query"
          description: "How many usernames to fetch"
          required: false
          type: "integer"
        - name: "page"
          in: "query"
          description: "Which page to fetch"
          required: false
          type: "integer"
      responses:
        200:
          description: "Users successfully fetched"
          schema:
            type: "array"
            items:
              type: "string"
        500:
          description: "Internal Server Error"
  /users/{username}:
    patch:
      tags:
        - "users"
      summary: "Edits a user"
      parameters:
        - in: "header"
          name: "Authorization"
          description: "JWT Bearer Access Token"
          required: true
          type: "string"
        - name: "username"
          in: "path"
          description: "Username of the user to edit"
          required: true
          type: "string"
        - in: "body"
          name: "password"
          description: "New password to update"
          required: true
          schema:
            type: "object"
            properties:
              password:
                type: "string"
      responses:
        204:
          description: "User successfully edited"
        400:
          description: "Bad Request, see error message"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized request."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
    delete:
      tags:
        - "users"
      summary: "Removes a user"
      parameters:
        - in: "header"
          name: "Authorization"
          description: "JWT Bearer Access Token"
          required: true
          type: "string"
        - name: "username"
          in: "path"
          description: "Username of the user to delete"
          required: true
          type: "string"
      responses:
        204:
          description: "User successfully removed"
        400:
          description: "Bad Request, see error message"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized request."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"

  # ============== SESSIONS ==============
...
  /votes/{memeId}:
    put:
      tags:
        - "votes"
      summary: "Vote on a meme"
      parameters:
        - in: "header"
          name: "Authorization"
          description: "JWT Bearer Access Token"
          required: true
          type: "string"
        - name: "memeId"
          in: "path"
          description: "ID of the meme"
          required: true
          type: "string"
        - name: "body"
          in: "body"
          description: "Username of the signed in user and vote to add"
          required: true
          schema:
            $ref: "#/definitions/NewVote"
      responses:
        200:
          description: "Vote successfully posted"
          schema:
            $ref: "#/definitions/Vote"
        400:
          description: "Bad Request, see error message"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized request."
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
    get:
      tags:
        - "votes"
      summary: "Get a user's vote on a meme"
      parameters:
        - name: "memeId"
          in: "path"
          description: "ID of the meme"
          required: true
          type: "integer"
        - name: "body"
          in: "body"
          description: "Name of the user"
          required: true
          schema:
            type: "object"
            properties:
              username:
                type: "string"
      responses:
        200:
          description: "Vote successfully fetched"
          schema:
            $ref: "#/definitions/Vote"
        400:
          description: "Bad Request, see error message"
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
    delete:
      tags:
        - "votes"
      summary: "Undo a vote on a meme"
      parameters:
        - in: "header"
          name: "Authorization"
          description: "JWT Bearer Access Token"
          required: true
          type: "string"
        - name: "memeId"
          in: "path"
          description: "ID of the meme"
          required: true
          type: "integer"
        - name: "body"
          in: "body"
          description: "Name of the user"
          required: true
          schema:
            type: "object"
            properties:
              username:
                type: "string"
      responses:
        204:
          description: "Vote successfully removed"
        400:
          description: "Bad Request, see error message"
          schema:
            $ref: "#/definitions/Error"
        401:
          description: "Unauthorized request."
          schema:
            $ref: "#/definitions/Error"
        404:
          description: "Not Found"
          schema:
            $ref: "#/definitions/Error"
        500:
          description: "Internal Server Error"
definitions:
  Error:
    type: "object"
    required:
      - "error"
    properties:
      error:
        type: "string"
  User:
    type: "object"
    required:
      - "username"
      - "password"
    properties:
      username:
        type: "string"
      password:
        type: "string"
...
  Vote:
    type: "object"
    required:
      - "memeId"
      - "username"
      - "vote"
    properties:
      memeId:
        type: "integer"
      username:
        type: "string"
      vote:
        type: "integer"
  NewVote:
    type: "object"
    required:
      - "username"
      - "vote"
    properties:
      username:
        type: "string"
      vote:
        type: "integer"
...